Download script improvements

+++ scripts/wrangling.R
@@ -24,13 +24,13 @@ set.seed(1234)
 
 opt <- docopt(doc)
 
-main <- function(data_path, train_path, test_path, target, remove_outliers, train_size){
+main <- function(data_path, train_path, test_path, target, remove_outliers, train_size) {
   data <- load(data_path)
   wrangled_data <- wrangling(data, target, remove_outliers)
   list_traintest <- split_data(wrangled_data, train_size)
-  
+
   print_dataset(train_path, list_traintest[[1]])
-  print_dataset(test_path, list_traintest[[1]])
+  print_dataset(test_path, list_traintest[[2]])
 }
 
 #' Loads the data from a provided path
@@ -40,10 +40,10 @@ main <- function(data_path, train_path, test_path, target, remove_outliers, trai
 #' @examples
 #' load('../data/vehicles.csv')
 load <- function(data_path) {
-  if (file.exists(data_path)){
+  if (file.exists(data_path)) {
     return(readr::read_csv(data_path) %>% select(-c(description, county, url, region_url)))
   }
-  else{
+  else {
     print("The path does not exist")
   }
 }
@@ -59,14 +59,14 @@ load <- function(data_path) {
 #' @examples
 #' wrangling(vehicles, price, TRUE)
 wrangling <- function(data, target, remove_outliers) {
-  for(i in names(data)){
-    if(is.character(data[[i]])){
+  for (i in names(data)) {
+    if (is.character(data[[i]])) {
       data[[i]] <- if_else(is.na(data[[i]]), "No value", data[[i]])
     }
   }
-  
-  if (remove_outliers == "YES"){
-    data_filtered <- data[target <= quantile(data[[target]],c(0.99)),]
+
+  if (remove_outliers == "YES") {
+    data_filtered <- data[data[[target]] <= quantile(data[[target]], c(0.99)),]
   } else {
     data_filtered <- data
   }
@@ -80,19 +80,19 @@ wrangling <- function(data, target, remove_outliers) {
 #' @return the data with a split column TRUE for train dataset FALSE test
 #' @examples
 #' split_data(vehicles, 0.9)
-split_data <- function(data, train_size){
+split_data <- function(data, train_size) {
   train_size <- as.double(train_size)
-  if(train_size >= 0 && train_size <= 1){
+  if (train_size >= 0 && train_size <= 1) {
     sample_size <- floor(as.double(train_size) * nrow(data))
     train_id <- sample(seq_len(nrow(data)), size = sample_size)
-    
-    train <- data[train_id, ]
-    test <- data[-train_id, ]
-    
-    list_results <- list(train,test)
+
+    train <- data[train_id,]
+    test <- data[-train_id,]
+
+    list_results <- list(train, test)
     return(list_results)
   }
-  else{
+  else {
     print("The train_size parameter is not valid")
   }
 }
@@ -100,16 +100,17 @@ split_data <- function(data, train_size){
 
 #' Prints the data the a provided paths
 #'
-#' @param data_path path to print the data
-#' @param data dataset to print
+#' @param data_path Path to print the data
+#' @param data Dataset to print
+#' @param replace Replace existing file or stop
 #' @return Nothing
 #' @examples
 #' print_train_test('../data/vehicles_train.csv',vehicles_train)
-print_dataset <- function(data_path, data) {
-  if (file.exists(data_path)){
+print_dataset <- function(data_path, data, replace = TRUE) {
+  if (file.exists(data_path) && !replace) {
     print("The file already exists")
   }
-  else{
+  else {
     write_csv(data, data_path)
   }
 }
